# api/openapi.json
{"openapi":"3.0.0","info":{"title":"iot-demo","description":"","version":"1.0.0"},"servers":[{"url":"https://localhost:5100/api"}],"tags":[{"name":"pressure"},{"name":"users"},{"name":"temperature"},{"name":"humidity"}],"paths":{"/pressure":{"get":{"summary":"","tags":["pressure"],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"boolean"}}}}}},"post":{"summary":"","tags":["pressure"],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"boolean"}}}}},"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"value":{"type":"string"},"addedDate":{"type":"string"}}}}}}},"put":{"summary":"","tags":["pressure"],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"boolean"}}}}},"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"value":{"type":"string"},"addedDate":{"type":"string"},"id":{"type":"string"}}}}}}},"delete":{"summary":"","tags":["pressure"],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"boolean"}}}}},"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"string"}}}}}}}},"/pressure/{filter}":{"get":{"summary":"","tags":["pressure"],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"boolean"}}}}},"parameters":[{"name":"filter","in":"path","required":true,"schema":{"type":"string"}}]}},"/pressure/pagination":{"get":{"summary":"pagination","tags":["pressure"],"responses":{"200":{"description":"pagination","content":{"application/json":{"schema":{"type":"boolean"}}}}},"parameters":[{"name":"pageSize","in":"query","schema":{"type":"integer"}},{"name":"page","in":"query","schema":{"type":"integer"}}]}},"/users/register":{"post":{"summary":"register","tags":["users"],"responses":{"200":{"description":"register","content":{"application/json":{"schema":{"type":"boolean"}}}}},"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"gender":{"type":"string"},"phoneNumber":{"type":"string"},"city":{"type":"string"},"country":{"type":"string"},"userType":{"type":"string"}}}}}}}},"/users/edit":{"put":{"summary":"edit","tags":["users"],"responses":{"200":{"description":"edit","content":{"application/json":{"schema":{"type":"boolean"}}}}},"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"gender":{"type":"string"},"phoneNumber":{"type":"string"},"city":{"type":"string"},"country":{"type":"string"},"userType":{"type":"string"},"id":{"type":"string"}}}}}}}},"/users/delete":{"delete":{"summary":"delete","tags":["users"],"responses":{"200":{"description":"delete","content":{"application/json":{"schema":{"type":"boolean"}}}}},"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"string"}}}}}}}},"/users/loginsingle":{"post":{"summary":"loginsingle","tags":["users"],"responses":{"200":{"description":"loginsingle","content":{"application/json":{"schema":{"type":"boolean"}}}}},"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}}}}}}},"/users/loginmulti":{"post":{"summary":"loginmulti","tags":["users"],"responses":{"200":{"description":"loginmulti","content":{"application/json":{"schema":{"type":"boolean"}}}}},"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}}}}}}},"/users/checksession":{"post":{"summary":"checksession","tags":["users"],"responses":{"200":{"description":"checksession","content":{"application/json":{"schema":{"type":"boolean"}}}}},"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{}}}}}}},"/users/forgotpassword":{"post":{"summary":"forgotpassword","tags":["users"],"responses":{"200":{"description":"forgotpassword","content":{"application/json":{"schema":{"type":"boolean"}}}}},"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"email":{"type":"string"}}}}}}}},"/users/resetpassword":{"post":{"summary":"resetpassword","tags":["users"],"responses":{"200":{"description":"resetpassword","content":{"application/json":{"schema":{"type":"boolean"}}}}},"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{}}}}}}},"/temperature":{"get":{"summary":"","tags":["temperature"],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"boolean"}}}}}},"post":{"summary":"","tags":["temperature"],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"boolean"}}}}},"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"value":{"type":"string"},"addedDate":{"type":"number"}}}}}}},"put":{"summary":"","tags":["temperature"],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"boolean"}}}}},"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"value":{"type":"string"},"addedDate":{"type":"number"},"id":{"type":"string"}}}}}}},"delete":{"summary":"","tags":["temperature"],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"boolean"}}}}},"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"string"}}}}}}}},"/temperature/{filter}":{"get":{"summary":"","tags":["temperature"],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"boolean"}}}}},"parameters":[{"name":"filter","in":"path","required":true,"schema":{"type":"string"}}]}},"/temperature/pagination":{"get":{"summary":"pagination","tags":["temperature"],"responses":{"200":{"description":"pagination","content":{"application/json":{"schema":{"type":"boolean"}}}}},"parameters":[{"name":"pageSize","in":"query","schema":{"type":"integer"}},{"name":"page","in":"query","schema":{"type":"integer"}}]}},"/humidity":{"get":{"summary":"","tags":["humidity"],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"boolean"}}}}}},"post":{"summary":"","tags":["humidity"],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"boolean"}}}}},"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"value":{"type":"string"},"addedDate":{"type":"number"}}}}}}},"put":{"summary":"","tags":["humidity"],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"boolean"}}}}},"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"value":{"type":"string"},"addedDate":{"type":"number"},"id":{"type":"string"}}}}}}},"delete":{"summary":"","tags":["humidity"],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"boolean"}}}}},"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"string"}}}}}}}},"/humidity/{filter}":{"get":{"summary":"","tags":["humidity"],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"boolean"}}}}},"parameters":[{"name":"filter","in":"path","required":true,"schema":{"type":"string"}}]}},"/humidity/pagination":{"get":{"summary":"pagination","tags":["humidity"],"responses":{"200":{"description":"pagination","content":{"application/json":{"schema":{"type":"boolean"}}}}},"parameters":[{"name":"pageSize","in":"query","schema":{"type":"integer"}},{"name":"page","in":"query","schema":{"type":"integer"}}]}}}}